Introduce IndexingHandler to avoid circularity
