+++ manager/src/main/java/io/gingersnapproject/search/QueryHandler.java
@@ -4,8 +4,6 @@ import javax.inject.Inject;
 import javax.inject.Singleton;
 
 import io.gingersnapproject.Caches;
-import io.gingersnapproject.configuration.Configuration;
-import io.gingersnapproject.configuration.Rule;
 import io.smallrye.mutiny.Multi;
 import io.smallrye.mutiny.Uni;
 
@@ -15,24 +13,6 @@ public class QueryHandler {
    Caches caches;
    @Inject
    SearchBackend searchBackend;
-   @Inject
-   Configuration configuration;
-
-   public Uni<String> put(String indexName, String documentId, String jsonString) {
-      Rule rule = configuration.rules().get(indexName);
-      if (rule == null || !rule.queryEnabled()) {
-         return Uni.createFrom().nullItem();
-      }
-      return searchBackend.put(indexName, documentId, jsonString);
-   }
-
-   public Uni<String> remove(String indexName, String documentId) {
-      Rule rule = configuration.rules().get(indexName);
-      if (rule == null || !rule.queryEnabled()) {
-         return Uni.createFrom().nullItem();
-      }
-      return searchBackend.remove(indexName, documentId);
-   }
 
    public Multi<String> query(String query) {
       Uni<SearchResult> result = searchBackend.query(query);
